Fixed cosmetic problem with "Seconds" modes and Torch mode by doing gamma correction
Even with previous commit, the Torch mode was difficult to see for low luminosity values.
"Seconds" modes make usage of low luminosity values and are also affected.

The solution is to apply a gamma correction of 5, resulting in a steep curve right from
low values (almost an offset in practice), and better visibility of luminosity changes.